Update redis-py to v6 (#5)
`redis-py` is currently pinned to <v6, which is unnecessary afaict,
since the breaking changes that v6 introduces are pretty minor. This PR
bumps the version to v6.4.0. redis-py v5 is pretty buggy, and has some
nasty failure modes around topology rediscovery, so it'll be good to get
off of it.